i hate pushing

+++ audit_model.py
@@ -947,7 +947,7 @@ def main():
     elif args.target_type == 'empty_sequence':
         # sequence length (same as existing chunks)
         seq_len = X_out.shape[1]
-        target_X = torch.zeros((1, seq_len), dtype=torch.float32)
+        target_X = torch.zeros((1, seq_len), dtype=torch.long)
         target_y = torch.tensor([9], dtype=torch.long)
     elif os.path.exists(args.target_type):
         # pre-crafted target sample
